Copy the apply and filter methods from groupby to separate files
